move the copyTask in to PrepareSandboxTask
